Created datasetloader for SignToGloss
