+++ models/signtogloss/datasetloader.py
@@ -1,0 +1,50 @@
+import json
+from sklearn.model_selection import train_test_split
+import torch
+from torch.utils.data import Dataset, DataLoader
+from torch.nn.utils.rnn import pad_sequence, pack_padded_sequence
+
+class SignToVideoDataset(Dataset):
+    def __init__(self, glosses, features):
+        super().__init__()
+        self.glosses = glosses
+        self.features = features
+        
+    def __len__(self):
+        return len(self.glosses)
+    
+    def __getitem__(self, index):
+        return self.glosses[index], torch.tensor(self.features[index])
+
+def collate_fn(batch):
+    glosses, features = zip(*batch)
+    seq_len = [len(feature) for feature in features]
+    
+    padded = pad_sequence(features, batch_first=True)
+    return glosses, padded, seq_len
+
+def load_data(batch_size, random_state=29, test_size=0.1):
+    with open('wsasl.json', 'r') as file:
+        data = json.load(file)
+    
+    data = [(sample["label"], sample["features"]) for sample in data]
+    glosses, features = zip(*data)
+
+    dataset = SignToVideoDataset(glosses, features)
+
+    gloss_train, gloss_test, features_train, features_test = \
+    train_test_split(glosses, features, random_state=29, test_size=test_size, shuffle=True)
+
+    train = SignToVideoDataset(gloss_train, features_train)
+    train_loader = DataLoader(train, batch_size=batch_size, collate_fn=collate_fn, shuffle=True)
+
+    return train_loader
+
+train_loader = load_data(5)
+
+for gloss, feature, seq_len in train_loader:
+    print(seq_len)
+    print(feature.shape)
+
+    packed = pack_padded_sequence(feature, seq_len, batch_first=True, enforce_sorted=False)
+    print(packed)
